replace legacy wallet configs with new source
