+++ packages/connectkit/src/components/Pages/Connectors/styles.ts
@@ -266,67 +266,3 @@ export const ConnectorIcon = styled(motion.div)`
     height: 100%;
   }
 `;
-export const MobileConnectorsContainer = styled(motion.div)`
-  display: flex;
-  flex-direction: row;
-  align-items: flex-start;
-  justify-content: space-between;
-  padding: 14px 0 28px;
-  margin: 0 0;
-`;
-
-export const MobileConnectorButton = styled(motion.button)`
-  --background: var(--ck-body-background-secondary);
-  cursor: pointer;
-  user-select: none;
-  position: relative;
-  padding: 0;
-  width: 100%;
-  min-width: 25%;
-  font-size: 13px;
-  font-weight: 500;
-  line-height: 13px;
-  text-align: center;
-  transition: 100ms ease;
-  transition-property: transform, opacity;
-
-  background: none;
-
-  &:disabled {
-    cursor: not-allowed;
-    opacity: 0.4;
-  }
-
-  &:not(:disabled) {
-    &:active {
-      transform: scale(0.97);
-    }
-  }
-`;
-
-export const MobileConnectorLabel = styled(motion.span)`
-  display: block;
-  padding: 10px 0 0;
-  color: var(--ck-body-color);
-  opacity: 0.75;
-  overflow: hidden;
-  white-space: nowrap;
-  text-overflow: ellipsis;
-`;
-
-export const MobileConnectorIcon = styled(motion.div)`
-  position: relative;
-  margin: 0 auto;
-  width: 60px;
-  height: 60px;
-  overflow: hidden;
-  svg {
-    display: block;
-    position: relative;
-    overflow: hidden;
-    border-radius: 27.5%;
-    transform: translate3d(0, 0, 0);
-    width: 100%;
-    height: 100%;
-  }
-`;
