+++ packages/connectkit/src/components/Pages/DownloadApp/index.tsx
@@ -10,7 +10,7 @@ import CustomQRCode from '../../Common/CustomQRCode';
 
 import useLocales from '../../../hooks/useLocales';
 import { useContext } from '../../ConnectKit';
-import { useWallet } from '../../../hooks/useWallets';
+import { useWallet } from '../../../wallets/useWallets';
 
 const DownloadApp = () => {
   const context = useContext();
