--- packages/connectkit/src/wallets/connectors/brave.tsx
@@ -1,19 +1,0 @@
-import { LegacyWalletProps } from './../wallet';
-
-import Logos from './../../assets/logos';
-import { isBrave } from '../../utils/wallets';
-
-export const brave = (): LegacyWalletProps => {
-  const isInstalled = isBrave();
-
-  return {
-    id: 'brave',
-    name: 'Brave Wallet',
-    shortName: 'Brave',
-    icon: <Logos.Brave />,
-    scannable: false,
-    downloadUrls: {},
-    installed: isInstalled,
-    createUri: (uri: string) => uri,
-  };
-};
