--- packages/connectkit/src/wallets/connectors/injected.tsx
@@ -1,18 +1,0 @@
-import { LegacyWalletProps } from './../wallet';
-
-import { isMobile } from '../../utils';
-import Logos from './../../assets/logos';
-
-export const injected = (): LegacyWalletProps => {
-  const isInstalled = typeof window !== 'undefined' && Boolean(window.ethereum);
-  const shouldUseWalletConnect = isMobile() && !isInstalled;
-
-  return {
-    id: 'injected',
-    name: 'Browser Wallet',
-    shortName: 'browser',
-    scannable: false,
-    icon: <Logos.Injected />,
-    installed: Boolean(!shouldUseWalletConnect ? isInstalled : false),
-  };
-};
