--- packages/connectkit/src/wallets/connectors/walletConnect.tsx
@@ -1,13 +1,0 @@
-import { LegacyWalletProps } from './../wallet';
-
-import Logos from './../../assets/logos';
-
-export const walletConnect = (): LegacyWalletProps => {
-  return {
-    id: 'walletConnect',
-    name: 'Other Wallets',
-    icon: <Logos.WalletConnect />,
-    scannable: true,
-    createUri: (uri: string) => uri,
-  };
-};
