+++ packages/connectkit/src/wallets/walletConfigs.tsx
@@ -16,7 +16,7 @@ export type WalletConfigProps = {
   // Icon to use on the wallet list button. If not provided, `icon` will be used
   iconConnector?: React.ReactNode;
   // Defaults to `'circle'`, but some icons look better as squircle (e.g. if they have a background)
-  iconShape?: 'squircle' | 'circle';
+  iconShape?: 'squircle' | 'circle' | 'square';
   // Defaults to `false`, but some icons don't have a background and look better if they shrink to fit the container
   iconShouldShrink?: boolean;
   // Links to download the wallet
@@ -45,11 +45,18 @@ export type WalletConfigProps = {
 // Organised in alphabetical order by key
 export const walletConfigs: { [rdns: string]: WalletConfigProps } = {
   argent: {
+    name: 'Argent',
+    icon: <Logos.Argent />,
     downloadUrls: {
       download: 'https://connect.family.co/v0/download/argent',
       android:
         'https://play.google.com/store/apps/details?id=im.argent.contractwalletclient',
       ios: 'https://apps.apple.com/app/argent/id1358741926',
+    },
+    getWalletConnectDeeplink: (uri: string) => {
+      return isAndroid()
+        ? uri
+        : `https://argent.link/app/wc?uri=${encodeURIComponent(uri)}`;
     },
   },
   coinbaseWallet: {
@@ -65,6 +72,9 @@ export const walletConfigs: { [rdns: string]: WalletConfigProps } = {
       chrome:
         'https://chrome.google.com/webstore/detail/coinbase-wallet-extension/hnfanknocfeofbddgcijnmhnfnkdnaad',
     },
+    getWalletConnectDeeplink: (uri: string) => {
+      return `https://go.cb-w.com/wc?uri=${encodeURIComponent(uri)}`;
+    },
   },
   'com.coinbase.wallet': {
     name: 'Coinbase Wallet',
@@ -78,6 +88,9 @@ export const walletConfigs: { [rdns: string]: WalletConfigProps } = {
       ios: 'https://apps.apple.com/app/coinbase-wallet-store-crypto/id1278383455',
       chrome:
         'https://chrome.google.com/webstore/detail/coinbase-wallet-extension/hnfanknocfeofbddgcijnmhnfnkdnaad',
+    },
+    getWalletConnectDeeplink: (uri: string) => {
+      return `https://go.cb-w.com/wc?uri=${encodeURIComponent(uri)}`;
     },
   },
   coinbaseWalletSDK: {
@@ -93,13 +106,17 @@ export const walletConfigs: { [rdns: string]: WalletConfigProps } = {
       chrome:
         'https://chrome.google.com/webstore/detail/coinbase-wallet-extension/hnfanknocfeofbddgcijnmhnfnkdnaad',
     },
-    getWalletConnectDeeplink: (uri: string) => uri,
+    getWalletConnectDeeplink: (uri: string) => {
+      return `https://go.cb-w.com/wc?uri=${encodeURIComponent(uri)}`;
+    },
   },
   'com.crypto.wallet': {
     name: 'Crypto.com',
     shortName: 'Crypto',
   },
   dawn: {
+    name: 'Dawn Wallet',
+    shortName: 'Dawn',
     downloadUrls: {
       download:
         'https://apps.apple.com/us/app/dawn-ethereum-wallet/id1673143782',
@@ -110,11 +127,17 @@ export const walletConfigs: { [rdns: string]: WalletConfigProps } = {
   'co.family.wallet': {
     name: 'Family',
     shortName: 'Family',
+    icon: <Logos.Family />,
     iconShape: 'squircle',
     downloadUrls: {
       download: 'https://connect.family.co/v0/download/family',
       website: 'https://family.co',
       ios: 'https://family.co/download',
+    },
+    getWalletConnectDeeplink: (uri: string) => {
+      return isAndroid()
+        ? uri
+        : `familywallet://wc?uri=${encodeURIComponent(uri)}`;
     },
   },
   frame: {
@@ -130,10 +153,12 @@ export const walletConfigs: { [rdns: string]: WalletConfigProps } = {
       brave:
         'https://chrome.google.com/webstore/detail/frame-companion/ldcoohedfbjoobcadoglnnmmfbdlmmhf',
     },
+    getWalletConnectDeeplink: (uri: string) => uri,
   },
   frontier: {
     name: 'Frontier Wallet',
     shortName: 'Frontier',
+    icon: <Logos.Frontier />,
     downloadUrls: {
       download: 'https://connect.family.co/v0/download/frontier',
       ios: 'https://apps.apple.com/app/frontier-crypto-defi-wallet/id1482380988',
@@ -142,6 +167,9 @@ export const walletConfigs: { [rdns: string]: WalletConfigProps } = {
       website: 'https://frontier.xyz/',
       chrome:
         'https://chrome.google.com/webstore/detail/frontier-wallet/kppfdiipphfccemcignhifpjkapfbihd',
+    },
+    getWalletConnectDeeplink: (uri: string) => {
+      return isAndroid() ? uri : `frontier://wc?uri=${encodeURIComponent(uri)}`;
     },
   },
   injected: {
@@ -161,7 +189,7 @@ export const walletConfigs: { [rdns: string]: WalletConfigProps } = {
       ios: 'https://apps.apple.com/app/metamask/id1438144202',
       chrome:
         'https://chrome.google.com/webstore/detail/metamask/nkbihfbeogaeaoehlefnkodbefgpgknn',
-      firefox: 'https://addons.mozilla.org/en-US/firefox/addon/ether-metamask/',
+      firefox: 'https://addons.mozilla.org/firefox/addon/ether-metamask/',
       brave:
         'https://chrome.google.com/webstore/detail/metamask/nkbihfbeogaeaoehlefnkodbefgpgknn',
       edge: 'https://microsoftedge.microsoft.com/addons/detail/metamask/ejbalbakoplchlghecdalmeeeajnimhm',
@@ -179,7 +207,7 @@ export const walletConfigs: { [rdns: string]: WalletConfigProps } = {
       ios: 'https://apps.apple.com/app/metamask/id1438144202',
       chrome:
         'https://chrome.google.com/webstore/detail/metamask/nkbihfbeogaeaoehlefnkodbefgpgknn',
-      firefox: 'https://addons.mozilla.org/en-US/firefox/addon/ether-metamask/',
+      firefox: 'https://addons.mozilla.org/firefox/addon/ether-metamask/',
       brave:
         'https://chrome.google.com/webstore/detail/metamask/nkbihfbeogaeaoehlefnkodbefgpgknn',
       edge: 'https://microsoftedge.microsoft.com/addons/detail/metamask/ejbalbakoplchlghecdalmeeeajnimhm',
@@ -197,10 +225,15 @@ export const walletConfigs: { [rdns: string]: WalletConfigProps } = {
       ios: 'https://apps.apple.com/app/metamask/id1438144202',
       chrome:
         'https://chrome.google.com/webstore/detail/metamask/nkbihfbeogaeaoehlefnkodbefgpgknn',
-      firefox: 'https://addons.mozilla.org/en-US/firefox/addon/ether-metamask/',
+      firefox: 'https://addons.mozilla.org/firefox/addon/ether-metamask/',
       brave:
         'https://chrome.google.com/webstore/detail/metamask/nkbihfbeogaeaoehlefnkodbefgpgknn',
       edge: 'https://microsoftedge.microsoft.com/addons/detail/metamask/ejbalbakoplchlghecdalmeeeajnimhm',
+    },
+    getWalletConnectDeeplink: (uri: string) => {
+      return isAndroid()
+        ? uri
+        : `https://metamask.app.link/wc?uri=${encodeURIComponent(uri)}`;
     },
   },
   'app.phantom': {
@@ -210,6 +243,7 @@ export const walletConfigs: { [rdns: string]: WalletConfigProps } = {
   'me.rainbow': {
     name: 'Rainbow Wallet',
     shortName: 'Rainbow',
+    icon: <Logos.Rainbow />,
     iconShape: 'squircle',
     downloadUrls: {
       download: 'https://connect.family.co/v0/download/rainbow',
@@ -243,9 +277,14 @@ export const walletConfigs: { [rdns: string]: WalletConfigProps } = {
     icon: <Logos.Safe />,
     downloadUrls: {
       download: 'https://connect.family.co/v0/download/safe',
-      website: 'https://safe.global/wallet',
-      android: 'https://play.google.com/store/apps/details?id=io.gnosis.safe',
+      website: 'https://safe.global/',
       ios: 'https://apps.apple.com/app/id1515759131',
+      android: 'https://play.google.com/store/apps/details?id=io.gnosis.safe',
+    },
+    getWalletConnectDeeplink: (uri: string) => {
+      return isAndroid()
+        ? uri
+        : `https://gnosis-safe.io/wc?uri=${encodeURIComponent(uri)}`;
     },
   },
   'xyz.talisman': {
@@ -264,6 +303,7 @@ export const walletConfigs: { [rdns: string]: WalletConfigProps } = {
   'com.trustwallet.app': {
     name: 'Trust Wallet',
     shortName: 'Trust',
+    icon: <Logos.Trust />,
     iconShouldShrink: true,
     downloadUrls: {
       download: 'https://connect.family.co/v0/download/trust',
@@ -271,12 +311,164 @@ export const walletConfigs: { [rdns: string]: WalletConfigProps } = {
         'https://play.google.com/store/apps/details?id=com.wallet.crypto.trustapp',
       ios: 'https://apps.apple.com/app/trust-crypto-bitcoin-wallet/id1288339409',
     },
+    getWalletConnectDeeplink(uri) {
+      return isAndroid()
+        ? uri
+        : `https://link.trustwallet.com/wc?uri=${encodeURIComponent(uri)}`;
+    },
+  },
+  infinityWallet: {
+    name: 'Infinity Wallet',
+    icon: <Logos.InfinityWallet />,
+    downloadUrls: {
+      download: 'https://connect.family.co/v0/download/infinityWallet',
+      website: 'https://infinitywallet.io/download',
+      chrome: 'https://infinitywallet.io/download',
+      firefox: 'https://infinitywallet.io/download',
+      brave: 'https://infinitywallet.io/download',
+      edge: 'https://infinitywallet.io/download',
+    },
+  },
+  imToken: {
+    name: 'imToken',
+    icon: <Logos.ImToken />,
+    downloadUrls: {
+      //website: 'https://support.token.im/hc/en-us/categories/360000925393',
+      download: 'https://connect.family.co/v0/download/imToken',
+      android: 'https://play.google.com/store/apps/details?id=im.token.app',
+      ios: 'https://itunes.apple.com/us/app/imtoken2/id1384798940',
+    },
+    getWalletConnectDeeplink: (uri: string) => {
+      return `imtokenv2://wc?uri=${encodeURIComponent(uri)}`;
+    },
+  },
+  unstoppable: {
+    name: 'Unstoppable',
+    icon: <Logos.Unstoppable />,
+    downloadUrls: {
+      download: 'https://connect.family.co/v0/download/unstoppable',
+      ios: 'https://apps.apple.com/app/bank-bitcoin-wallet/id1447619907',
+      android:
+        'https://play.google.com/store/apps/details?id=io.horizontalsystems.bankwallet',
+    },
+    getWalletConnectDeeplink: (uri: string) => {
+      return isAndroid()
+        ? uri
+        : `https://unstoppable.money/wc?uri=${encodeURIComponent(uri)}`;
+    },
+  },
+  onto: {
+    name: 'ONTO',
+    icon: <Logos.ONTO />,
+    downloadUrls: {
+      download: 'https://connect.family.co/v0/download/onto',
+      ios: 'https://apps.apple.com/app/onto-an-ontology-dapp/id1436009823',
+      android:
+        'https://play.google.com/store/apps/details?id=com.github.ontio.onto',
+      website: 'https://onto.app/en/download/',
+    },
+    getWalletConnectDeeplink: (uri: string) => {
+      return isAndroid()
+        ? uri
+        : `https://onto.app/wc?uri=${encodeURIComponent(uri)}`;
+    },
+  },
+  steak: {
+    name: 'Steak',
+    icon: <Logos.Steak />,
+    downloadUrls: {
+      download: 'https://connect.family.co/v0/download/steak',
+      android:
+        'https://play.google.com/store/apps/details?id=fi.steakwallet.app',
+      ios: 'https://apps.apple.com/app/steakwallet/id1569375204',
+      website: 'https://steakwallet.fi/download',
+    },
+    getWalletConnectDeeplink: (uri: string) => {
+      return isAndroid()
+        ? uri
+        : `https://links.steakwallet.fi/wc?uri=${encodeURIComponent(uri)}`;
+    },
+  },
+  ledger: {
+    name: 'Ledger Live',
+    shortName: 'Ledger',
+    icon: <Logos.Ledger />,
+    downloadUrls: {
+      download: 'https://connect.family.co/v0/download/ledger',
+      website: 'https://www.ledger.com/ledger-live/download#download-device-2',
+      android: 'https://play.google.com/store/apps/details?id=com.ledger.live',
+      ios: 'https://apps.apple.com/app/ledger-live-web3-wallet/id1361671700',
+    },
+    getWalletConnectDeeplink: (uri: string) => {
+      return isAndroid()
+        ? uri
+        : `ledgerlive://wc?uri=${encodeURIComponent(uri)}`;
+    },
+  },
+  zerion: {
+    name: 'Zerion',
+    icon: <Logos.Zerion />,
+    downloadUrls: {
+      download: 'https://connect.family.co/v0/download/zerion',
+      ios: 'https://apps.apple.com/app/apple-store/id1456732565',
+      android:
+        'https://play.google.com/store/apps/details?id=io.zerion.android',
+      website: 'https://zerion.io/',
+    },
+    getWalletConnectDeeplink: (uri: string) => {
+      return isAndroid()
+        ? uri
+        : `https://app.zerion.io/wc?uri=${encodeURIComponent(uri)}`;
+    },
+  },
+  slope: {
+    name: 'Slope',
+    icon: <Logos.Slope />,
+    downloadUrls: {
+      download: 'https://connect.family.co/v0/download/slope',
+      ios: 'https://apps.apple.com/app/slope-wallet/id1574624530',
+      android: 'https://play.google.com/store/apps/details?id=com.wd.wallet',
+      chrome:
+        'https://chrome.google.com/webstore/detail/slope-wallet/pocmplpaccanhmnllbbkpgfliimjljgo',
+      website: 'https://slope.finance/',
+    },
+    getWalletConnectDeeplink: (uri: string) => {
+      return isAndroid()
+        ? uri
+        : `https://slope.finance/app/wc?uri=${encodeURIComponent(uri)}`;
+    },
+  },
+  tokenPocket: {
+    name: 'TokenPocket Wallet',
+    icon: <Logos.TokenPocket />,
+    downloadUrls: {
+      website: 'https://www.tokenpocket.pro/en/download/app',
+      download: 'https://www.tokenpocket.pro/en/download/app',
+      android:
+        'https://play.google.com/store/apps/details?id=vip.mytokenpocket',
+      ios: 'https://apps.apple.com/us/app/tp-global-wallet/id6444625622',
+      chrome:
+        'https://chrome.google.com/webstore/detail/tokenpocket/mfgccjchihfkkindfppnaooecgfneiii',
+    },
+  },
+  talisman: {
+    name: 'Talisman',
+    icon: <Logos.Talisman />,
+    downloadUrls: {
+      download: 'https://talisman.xyz/download',
+      website: 'https://talisman.xyz',
+      chrome:
+        'https://chrome.google.com/webstore/detail/talisman-polkadot-wallet/fijngjgcjhjmmpcmkeiomlglpeiijkld',
+      firefox:
+        'https://addons.mozilla.org/en-US/firefox/addon/talisman-wallet-extension/',
+    },
   },
   walletConnect: {
     name: 'Other Wallets',
     shortName: 'Other',
     icon: <Logos.WalletConnect background />,
     iconConnector: <Logos.OtherWallets />,
+    iconShape: 'square',
     getWalletConnectDeeplink: (uri: string) => uri,
   },
 };
